Added updating of particles functionality
Added member function in order to update particle positions and velocities using calculated gradient of potential field. void update_particles()

Added two test cases to test this. The first takes 10 timesteps and checks to see if the distance between two separated particles consistently decreases.
The second takes 200,000 timesteps and checks to see if average distance between the particles is within 0.1 of 0. It also does this for the respecitve velocity.